requirements.txt
streamlit
requests
beautifulsoup4
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client